refactor : GPT 응답 결과 mypage로 이동 (#28)
