+++ src/docs/mypageDocs.ts
@@ -254,5 +254,111 @@ export const mypageSwaggerDocs = {
         },
       },
     },
+
+    "/api/mypage/profiles/recommendations": {
+      get: {
+        summary: "GPT 추천 결과 목록 조회",
+        description: "사용자가 저장한 GPT 추천 결과 리스트 반환",
+        tags: ["Mypage"],
+        security: [{ bearerAuth: [] }],
+        responses: {
+          200: {
+            description: "추천 결과 조회 성공",
+            content: {
+              "application/json": {
+                schema: {
+                  type: "array",
+                  items: {
+                    type: "object",
+                    properties: {
+                      _id: {
+                        type: "string",
+                        example: "660f62c89abf1b001c66e678",
+                      },
+                      profile: {
+                        type: "object",
+                        properties: {
+                          education: {
+                            type: "string",
+                            example: "대학교 응용수학과 학사",
+                          },
+                          experience: {
+                            type: "string",
+                            example: "학원 수학 선생님 아르바이트 3년",
+                          },
+                          skills: {
+                            type: "array",
+                            items: { type: "string" },
+                            example: [
+                              "초,중,고등 수학 지도 가능",
+                              "토익 800점 이상",
+                              "학점 4.0이상",
+                            ],
+                          },
+                          languages: {
+                            type: "array",
+                            items: { type: "string" },
+                            example: ["한국어", "일본어"],
+                          },
+                          desiredSalary: { type: "number", example: 5000 },
+                          desiredJob: {
+                            type: "string",
+                            example: "없음",
+                          },
+                          additionalNotes: {
+                            type: "string",
+                            example: "아시아 국가 희망",
+                          },
+                        },
+                      },
+                      rankings: {
+                        type: "array",
+                        items: {
+                          type: "object",
+                          properties: {
+                            country: { type: "string" },
+                            job: { type: "string" },
+                            reason: { type: "string" },
+                          },
+                        },
+                        example: [
+                          {
+                            country: "일본",
+                            job: "수학 교사",
+                            reason:
+                              "일본어 구사 가능하며, 학원에서의 수학 지도 경험이 있음",
+                          },
+                          {
+                            country: "싱가포르",
+                            job: "수학 튜터",
+                            reason:
+                              "싱가포르는 교육열이 높아 수학 튜터에 대한 수요가 많음",
+                          },
+                          {
+                            country: "중국",
+                            job: "국제학교 수학 선생님",
+                            reason:
+                              "중국의 국제학교에서는 외국인 선생님을 선호하며, 토익 점수와 학점이 높아 유리함",
+                          },
+                        ],
+                      },
+
+                      createdAt: {
+                        type: "string",
+                        format: "date-time",
+                        example: "2025-04-04T05:34:21.201Z",
+                      },
+                    },
+                  },
+                },
+              },
+            },
+          },
+          401: { description: "인증 실패 (토큰 없음 또는 유효하지 않음)" },
+          404: { description: "저장된 추천 결과가 없습니다." },
+          500: { description: "서버 오류" },
+        },
+      },
+    },
   },
 };
